refactor: extract contract governance native bridge
