+++ docs/engineering_convergence/contract_governance_responsibility_map.md
@@ -3,7 +3,7 @@
 Date: 2026-07-13
 Owner: Platform owner
 Target file: `addons/smart_core/utils/contract_governance.py`
-Current size: 4,121 lines
+Current size: 3,932 lines
 Phase: staged responsibility split
 
 ## Purpose
@@ -37,7 +37,7 @@ and by a smaller public module layout.
 | imported + 595-670 | User surface normalization | Capability and scene normalization are delegated; scene sanitization, search/action noise reduction, and user-surface policies remain in facade/user-surface helpers. | `contract_governance_capabilities.py`, `contract_governance_scenes.py`, and `contract_governance_user_surface.py`. |
 | 1267-2178 | Project and enterprise governance | Scene list metadata, project form/list/kanban/task transforms, enterprise company/department/user forms. | `contract_governance_project_profiles.py` and `contract_governance_enterprise_profiles.py`. |
 | 1591-1965 + imported | Standard list governance | Standard list profile application, delegated toolbar labels, tier-review list shaping. | `contract_governance_list_surface.py` plus future standard-list policy module. |
-| 2599-3044 | Native surface and scene bridge | Visible-field access realignment, native surface normalization, scene contract v1 bridge, labels, relation semantics. | `contract_governance_native_bridge.py`. |
+| 1917-2290 + imported | Native surface and scene bridge | Visible-field access realignment remains in facade; native surface normalization, scene semantic surface normalization, scene contract v1 envelope, search/action bridge, labels, and relation semantics are staged. | `contract_governance_native_bridge.py` plus remaining label/relation helpers. |
 | 3047-4499 | Form policy and render semantics | Render profile, view capabilities, field groups, layout backfill, action policies, validation rules, create-profile noise hiding, canonical key mapping. | `contract_governance_form_policy.py`. |
 | 4502-4742 | Domain override and diagnostics | Domain override registry/application, diagnostics, snapshots, surface mapping. | `contract_governance_diagnostics.py`. |
 | 4745-4820 | Main pipeline | Orchestrates all projection transforms and attaches metadata. | Keep as thin facade in `contract_governance.py`. |
@@ -60,6 +60,7 @@ and by a smaller public module layout.
 | `contract_governance_capabilities_split_guard.py` | Capability normalization extraction compatibility, shared registry behavior, and purity lock. |
 | `contract_governance_scenes_split_guard.py` | Scene normalization extraction compatibility, shared semantic-profile registry behavior, and purity lock. |
 | `contract_governance_list_surface_split_guard.py` | Standard list toolbar label extraction compatibility and row-open action behavior lock. |
+| `contract_governance_native_bridge_split_guard.py` | Native/scene bridge extraction compatibility, envelope derivation, search surface, and action bridge behavior lock. |
 
 ## Extraction Order
 
@@ -193,3 +194,20 @@ Stage 6 is complete when:
 - the extracted module remains projection-only: no ORM calls, HTTP calls,
   routing, file IO, or environment access;
 - `contract_governance.py` is locked at `<=4121` lines for this stage.
+
+## Stage 7 Target
+
+Stage 7 is complete when:
+
+- `contract_governance_native_bridge.py` owns native and scene bridge normalization,
+  native-view contract surface normalization, scene semantic surface
+  normalization, search-surface derivation, scene action derivation, and scene
+  contract v1 envelope construction;
+- `contract_governance.py` keeps the previous private helper names as wrappers
+  so the main pipeline order and direct helper references remain stable;
+- visible-field access realignment, business labels, and relation-entry
+  semantics remain in the facade until their behavior has narrower fixture
+  coverage;
+- the extracted module remains projection-only: no ORM calls, HTTP calls,
+  routing, file IO, or environment access;
+- `contract_governance.py` is locked at `<=3932` lines for this stage.
