+++ make/ci.mk
@@ -693,6 +693,7 @@ ci.local.quick: guard.prod.forbid architecture.complexity_baseline_lock verify.u
 	@python3 scripts/verify/contract_governance_capabilities_split_guard.py
 	@python3 scripts/verify/contract_governance_scenes_split_guard.py
 	@python3 scripts/verify/contract_governance_list_surface_split_guard.py
+	@python3 scripts/verify/contract_governance_native_bridge_split_guard.py
 	@python3 scripts/verify/frontend_page_contract_boundary_guard.py
 	@python3 scripts/verify/frontend_page_contract_orchestration_consumption_guard.py
 	@python3 scripts/verify/frontend_contract_consumer_intrusion_guard.py
