+++ scripts/verify/contract_governance_responsibility_map_guard.py
@@ -47,6 +47,7 @@ REQUIRED_GUARDS = [
     "contract_governance_capabilities_split_guard.py",
     "contract_governance_scenes_split_guard.py",
     "contract_governance_list_surface_split_guard.py",
+    "contract_governance_native_bridge_split_guard.py",
 ]
 
 INVARIANTS = [
@@ -82,7 +83,7 @@ def main() -> int:
 
     for token in [
         "Contract Governance Responsibility Map",
-        "Current size: 4,121 lines",
+        "Current size: 3,932 lines",
         "projection-only",
         "Do not start mechanical extraction",
         "## Public Entry Points",
@@ -110,6 +111,9 @@ def main() -> int:
         "## Stage 6 Target",
         "`contract_governance_list_surface.py` owns standard list toolbar labels",
         "`contract_governance.py` is locked at `<=4121` lines",
+        "## Stage 7 Target",
+        "`contract_governance_native_bridge.py` owns native and scene bridge normalization",
+        "`contract_governance.py` is locked at `<=3932` lines",
     ]:
         if token not in doc:
             errors.append(f"responsibility map missing token: {token}")
